feat: made events page dynamic
The event page is dynamic now. Next steps include making Event Carousel dynamic and event/id page
